pyproject.toml
[tool.poetry]
name = "django-website"
version = "0.1.0"
description = ""
authors = ["wagner <wmaluk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
psycopg2 = "^2.9.3"
django-import-export = "^2.8.0"
python-decouple = "^3.6"
dj-static = "^0.0.6"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
